Fixes for re-enabling more MSVC level 4 warnings: net/ edition
This contains fixes for the following sorts of issues:
* Octal escape sequence terminated by decimal number
* Possibly-uninitialized local variable
* Assignment inside conditional
* Signedness mismatch

This also contains a very small number of other cleanups to nearby code.

BUG=81439
TEST=none
R=rch@chromium.org

Review URL: https://codereview.chromium.org/374033002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@281875 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 6547ccdb9c236e1c2b6fee20a73a8d691481bb88

+++ net/dns/host_resolver.h
@@ -120,7 +120,7 @@ class NET_EXPORT HostResolver {
   static const size_t kDefaultParallelism = 0;
 
   // Set Options.max_retry_attempts to this to select a default retry value.
-  static const size_t kDefaultRetryAttempts = -1;
+  static const size_t kDefaultRetryAttempts = static_cast<size_t>(-1);
 
   // If any completion callbacks are pending when the resolver is destroyed,
   // the host resolutions are cancelled, and the completion callbacks will not
